Add new SerialTx and SerialRx drivers
These drivers are operating on TX and RX of uart separately. The TX
driver implements UartWriter. The RX driver does not implement
UartReader, but instead expects the Device to implement an EventHandler
for incoming data. This allows unbuffered I/O with UART.

The echo server example is simplified and uses the separate tx and rx
drivers.