Porting knative-client example over to drogue-device
This revealed an issue with the initialization behavior of
drogue-device: specifically, it is incorrect to assume that on_start
will be triggered before all actors have been initialized. So the
attempt to join the network in on_start always failed when the app
started because the modem took longer to initialize.

Co-authored-by: Ulf Lilleengen <ulf.lilleengen@gmail.com>