Use button to join and avoid stm32f4-esp8266 on_start race condition
The latest release of stm32f4xx-hal didn't include check_interrupt()
in its 'ExtiPin' trait. And its master branch HEAD was incompatible
with the latest nucleo BSP, so I got rid of the nucleo dep to bring in
the latest HAL. This finally enabled the join to work in response to a
button press.