Split sx127x driver into actor and controller
Split driver into an "actor", "controller", and "interrupt" component.

The interrupt component receives interrupts and passes on the appropriate
event to the controller component.

The controller component receives messages from interrupt and timers and
provides responses to commands issues by the actor component.

The actor component forwards commands to the controller components and
awaits the state processing to complete before responding to the user.

Implement receive in LoRa API