Enable more than one socket to be opened
The take_while logic only worked for the first socket. Subsequent
opens all failed. And I don't think we can rely on the order either,
since the first socket opened may not be the first closed.

"This is probably a good place for a unit test!" he said, as he
commited the fix without a unit test.