Change actor API to use a single future and mailbox
Simplify actor API by removing on_message, and providing the Actor receive
channel to the actor in on_start. This allows an actor to perform select
on receive and other futures, thereby simplifying implementation for
actors that need periodic tasks to be run.

Simplify matrix actor and examples that need periodic time keeping.