Refactor configuration to support more config file locations
Configuration entries are pulled from the `.drogue/config.toml` found
beneath $HOME and the project manifest directory, typically
$PWD. Additionally, all the latter's parents will be
searched. Precendence is determined by the order of calls to
Config::merge(), so we start from the root and search down through the
manifest dir components i.e. last merge "wins". The
`.drogue/config.toml` beneath $HOME has the lowest precendence.

We also fail the compile if any entries are missing like we used
to. We can give a little better error message referencing
`.drogue/config.toml` and the missing value.

The next step is to create a proc macro that can query the
now-lazily-created static CONFIG hashmap and avoid all the file
inclusion of the files we're still creating from the CONFIG entries.