Refactor actor context and simplify actor configuration
This allows ActorContext to be created as statics and merges creation
and initialization of actors in a single place. This means actors no
longer have to define a configuration to use, but can simply add the
desired configuration at actor construction time. The actor instance itself is passed when mounting the actor context.

This allows actors to be used in more situations, and also removes the
need for DeviceContext.