Change led/app variables to have _address suffix
This commit suggests changing the name of the led variable to
led_address, and app to app_address, to (hopefully) improve
the readability of the code.

The motivation for this change was that I had to look up that
ActorContext's mount function returns an Address, and perhaps others
that are new to the code base will have the same issue.