Add BoardConfig type to Board trait
This commit adds an associated type named BoardConfig to the Board trait.
The intention of this is to allow a Board to be created with a
configuration object specific to that board by using the new_with_config
constructor function.

A default implementation is provided for new_with_config, for Boards that
don't have a configuration. For those boards that don't have a config the
unit type (()) is set as the BoardConfig value.

Refs: https://github.com/drogue-iot/drogue-device/issues/169