Change bootloader name to nucleo-wl55-bootloader
This commit changes the name of this crate to be
nucleo-wl55-bootloader instead of iot01a-bootloader.

The motivation for this change is that there another crate with the
same name in /examples/stm32l4/iot01a/bootloader.
And when building the EclipseCon 2022 Hackaton firmware the following
warning is generated:

$ cargo b --release
warning: skipping duplicate package `iot01a-bootloader` found at
`/.cargo/git/checkouts/drogue-device-5b9868033e74c452/05ef609/examples/stm32l4/iot01a/bootloader`

I've verified is by manually changing the name of the
nucleo-wl55-bootloader in the git checkouts directory and the warning
is gone. I looks like the dependency microbit-bsp is causing this and
I'm guessing here that all of the Cargo.toml files are parsed for this
dependency (git repo drogue-iot/drogue-device.git), and that the
workspace in drogue-device is not used in this case.